pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "“GustavodePaulaGorges”", email = "“gustavodep.gorges@gmail.com”"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "drf-spectacular>=0.26.4",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
]
